Function cpp_utils::dynamic_cast_mut

pub unsafe fn dynamic_cast_mut<R, T: DynamicCast<R>>(ptr: *mut T) -> *mut R

Converts type of a mutable pointer using DynamicCast implementation of the type. ptr must be either a null pointer or a valid pointer to an instance of T class or a class derived from T. Returns null pointer if ptr does not point to an instance of R. If ptr is null, this function does nothing and returns null pointer.